Init repo
This does not work at all right now but is taken out of another repo
where it was "almost" working how I would like. However I found that
the rules_oci has some missing features before I can completely move
over to using it. This repo will be here to pull out the parts I want
from the @distroless repo which should let me do 95% of what I want.

To note I have stolen debian_package_manager from the @distroless repo
however it is going to be re-written to work with any repo and not just
with debian. When I do it will be re-namespaced but for the initial
commit I want to leave it as it was originally.